feat: add asset management and model loading
- Introduced an `Assets` module for managing embedded assets.
- Added support for loading GLTF/GLB models and textures.
- Integrated a new `Model` struct for handling 3D models with materials.
- Updated the renderer to support drawing textured models.
- Added a new shader for rendering models with texture support.
- Included an example model (`AnisotropyBarnLamp.glb`) in the assets.
- Enhanced the camera and game state to accommodate new features.

+++ crates/client/src/game/mod.rs
@@ -10,7 +10,7 @@ use crate::render::Camera;
 
 const BASE_MOVE_SPEED: f32 = 3.0;
 const SPRINT_MULTIPLIER: f32 = 3.0;
-const MOUSE_SENSITIVITY: f32 = 0.0002;
+const MOUSE_SENSITIVITY: f64 = 0.0002;
 
 pub struct GameState {
     pub input: Input,
@@ -51,8 +51,8 @@ impl GameState {
         let (dx, dy) = self.input.consume_mouse_delta();
         if dx != 0.0 || dy != 0.0 {
             self.camera.rotate(
-                dx as f32 * MOUSE_SENSITIVITY,
-                -dy as f32 * MOUSE_SENSITIVITY,
+                dx * MOUSE_SENSITIVITY,
+                -dy * MOUSE_SENSITIVITY,
             );
         }
     }
